Enhancement #46: decompose math controller class
