+++ analysis_options.yaml
@@ -1,15 +1,4 @@
-include: package:solid_lints/analysis_options.yaml
-
+include: package:flutter_lints/flutter.yaml
 analyzer:
   errors:
-    always_use_package_imports: ignore
-    cast_nullable_to_non_nullable: ignore
     must_be_immutable: ignore
-    public_member_api_docs: ignore
-
-custom_lint:
-  rules: 
-    no_empty_block: false
-    cyclomatic_complexity: false
-    avoid_returning_widgets: false
-    avoid_non_null_assertion: false
