+++ lib/flutter_math_input.dart
@@ -1,5 +1,5 @@
 export 'package:provider/provider.dart';
-export 'keyboard_controller.dart';
+export 'math_controller.dart';
 export 'widgets/keyboard/basic_keyboard/basis_keyboard.dart';
 export 'widgets/keyboard/custom_math_icons.dart';
 export 'widgets/keyboard/keyboard.dart';
