+++ lib/widgets/keyboard/basic_keyboard/basic_keyboard_pages/greek_symbols_scroll.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../../../keyboard_controller.dart';
+import '../../../../math_controller.dart';
 import '../basis_keyboard.dart';
 
 const greekSymbolsList = [
