+++ lib/widgets/keyboard/basic_keyboard/basic_keyboard_pages/latin_alphabet_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../../../../keyboard_controller.dart';
+import '../../../../math_controller.dart';
 import '../../custom_math_icons.dart';
 import '../basis_keyboard.dart';
